Disallowed for moving viewport outside the map.
Also fixed the map to centered, when the viewport is bigger than the map.

Right now it looks as if this change was buggy, because you can go outside the map downwards and to the right, but that is just because the size is not set precisely enough in the map
data (it's not a bug, it's a feature).