Added Blood decals.
This commit introduces decals, in their basic form for now. It's enought for basic blood and other decals, but only drawn on the map. If we ever want decals on textures, a lot more
work needs to be done.

Changes:
- added RenderTexture for the map texture and made Graphics use it as the base texture,
- added DecalManager, abstract Decal and BloodDecal implementations,
- added data for the blood decal (images, mostly) and updated the README with the gfx sources,
- disabled ShotEffect.

Miss still remains to be done. It's not shown for now.